restructure the project and add NowPlayingJAX
